refactor: simplify debug symbols module dependencies
